Various fixes and improvements
- `experimentNumber` is now a `long`, according to the db model used by
experiments-manager;
- ant version in POM moved to a property;
- removed dot separators from dropwizard configurations, and changed to
more appropriate camelCase;
- api endpoint to generate a driver is now `/generatedriver`;
- `benchFlowComposePath` is now `benchFlowComposeAddress`;
- `generateConstraint` is now `generateBenchFlowFields`;
- improved regex for IP matching.

+++ api/src/main/java/cloud/benchflow/driversmaker/requests/Experiment.java
@@ -18,7 +18,7 @@ public class Experiment {
     private String benchmarkName;
 
     @JsonProperty("experimentNumber")
-    private int experimentNumber;
+    private long experimentNumber;
 
     @JsonProperty("trials")
     private int totalTrials;
@@ -31,11 +31,11 @@ public class Experiment {
         this.benchmarkName = benchmarkName;
     }
 
-    public int getExperimentNumber() {
+    public long getExperimentNumber() {
         return experimentNumber;
     }
 
-    public void setExperimentNumber(int experimentNumber) {
+    public void setExperimentNumber(long experimentNumber) {
         this.experimentNumber = experimentNumber;
     }
